Finally fixed WholeMLMover, LocalMLMover and FJC and SJF energies.
It turns out that (because rotations in 3D are not commutative), it makes
a difference where we start if we calculate the stat deviation.
Thus sum_of_stats and stat.is_similar_to (in forgi) are not suitable
to constrain sampling. Instead we use the difference between the actual stem
and the virtual stem that would be placed by the stat assigned to the broken ML segment, to constrain  our sampling.