Add a score-field to AMinorGeometry
The second column in the FR3D output contains a score.
The lower the score, the better the fit between query and target.
Scores for interactions are typically below 1.0, the FR3D default
is to count only matches below a score of 0.5

We use an artificial value of 1000 for non-interactions.